fix: move NodeRequestError to own module to avoid heavy transitive imports
