+++ src/request/node-request-error.ts
@@ -1,0 +1,12 @@
+import type {ClientRequest} from 'http'
+
+export class NodeRequestError extends Error {
+  request: ClientRequest
+  code?: string | undefined
+
+  constructor(err: NodeJS.ErrnoException, req: any) {
+    super(err.message)
+    this.request = req
+    this.code = err.code
+  }
+}
